refactoring

+++ controllers/PostController.js
@@ -1,9 +1,7 @@
-const Post = require('../models/Post');
-const PostAdapter = require('../adapters/Post');
 const PostService = require('../services/PostService');
 
 exports.createPost = async (req, res, next) => {
-    await PostService.createPost(req.user.id, req.body);
+    const createdPost = await PostService.createPost(req.user.id, req.body);
 
     res.status(201).json({ message: "Post created" });
 }
@@ -18,18 +16,33 @@ exports.getPosts = async (req, res) => {
     res.json(posts);
 
 }
+
 exports.getPost = async (req, res) => {
     const post = await PostService.getPost(req.params.id);
 
     res.json(post);
 }
-exports.updatePost = async (req, res) => {
-    const post = await PostService.updatePost(req.params.id, req.body);
 
-    res.json(post);
+exports.updatePost = async (req, res) => {
+    const post = await PostService.getPost(req.params.id);
+
+    if (post.owner.id.toString() !== req.user.id.toString()) {
+        return res.status(403).json({ error: 'You are not allowed to update this post' });
+    }
+
+    const updatePost = await PostService.updatePost(req.params.id, req.body);
+
+    res.json({ message: 'Post updated' });
 }
+
 exports.deletePost = async (req, res) => {
-    await PostService.deletePost(req.params.id);
+    const post = await PostService.getPost(req.params.id);
+
+    if (post.owner.id.toString() !== req.user.id.toString()) {
+        return res.status(403).json({ error: 'You are not allowed to update this post' });
+    }
+
+    const deletedPost = await PostService.deletePost(req.params.id);
 
     res.json({ message: 'Post deleted' });
 }
